refactor(spacing): improve name to index
